one bug and one change to message box
some bugs and changes to figure size and message box font size. Font size change required a change to a latex interpreter which necessitated a few other changes. The mode=1 functionality is currently not working because the date format is different and the extension is lower case instead of upper case in the attended mode data.